+++ selfdrive/car/hyundai/carcontroller.py
@@ -63,9 +63,7 @@ class CarController:
     self.send_lfa_mfa_lkas = CP.flags & HyundaiFlags.SEND_LFA.value
     self.last_button_frame = 0
     self.pcmCruiseButtonDelay = 0
-    self.dynamicJerk = 0
     self.jerkStartLimit = 1.0
-    self.jerkUpperLowerLimit = 12.0       
     self.speedCameraHapticEndFrame = 0
     self.hapticFeedbackWhenSpeedCamera = 0
     self.maxAngleFrames = MAX_ANGLE_FRAMES
@@ -131,9 +129,7 @@ class CarController:
 
     # *** common hyundai stuff ***
     if self.frame % 100 == 0:
-      self.dynamicJerk = int(Params().get("DynamicJerk", encoding="utf8"))
       self.jerkStartLimit = float(int(Params().get("JerkStartLimit", encoding="utf8"))) * 0.1
-      self.jerkUpperLowerLimit = float(int(Params().get("JerkUpperLowerLimit", encoding="utf8")))
       self.hapticFeedbackWhenSpeedCamera = int(Params().get("HapticFeedbackWhenSpeedCamera", encoding="utf8"))
       self.maxAngleFrames = int(Params().get("MaxAngleFrames", encoding="utf8"))
       self.softHoldMode = int(Params().get("SoftHoldMode", encoding="utf8"))
@@ -235,31 +231,22 @@ class CarController:
       if self.frame % 2 == 0 and self.CP.openpilotLongitudinalControl:
         # TODO: unclear if this is needed
         startingJerk = self.jerkStartLimit
+        jerkLimit = 5.0
         self.jerk_count += DT_CTRL
-        jerk_max = interp(self.jerk_count, [0, 1.5, 2.5], [startingJerk, startingJerk, self.jerkUpperLowerLimit])
+        jerk_max = interp(self.jerk_count, [0, 1.5, 2.5], [startingJerk, startingJerk, jerkLimit])
         a_diff = CS.out.aEgo - accel # (+)인경우 내려야, 
         speed_diff = CS.out.vEgo - actuators.speed
         if actuators.longControlState == LongCtrlState.off:
-          jerk_u = self.jerkUpperLowerLimit
-          jerk_l = self.jerkUpperLowerLimit
+          jerk_u = jerkLimit
+          jerk_l = jerkLimit
           self.jerk_count = 0
         elif actuators.longControlState == LongCtrlState.stopping or hud_control.softHold:
           jerk_u = 0.5
-          jerk_l = self.jerkUpperLowerLimit
+          jerk_l = jerkLimit
           self.jerk_count = 0
-        elif self.dynamicJerk == 1:
+        else:
           jerk_u = min(max(0.5, jerk * 2.0), jerk_max)
           jerk_l = min(max(1.0, -jerk * 2.0), jerk_max)
-        elif self.dynamicJerk == 3:
-          jerk_max_u = interp(a_diff, [-0.5, 0.5], [jerk_max, 0.5])
-          jerk_max_l = interp(a_diff, [-0.5, 0.5], [0.5, jerk_max])
-          jerk_u = interp(jerk, [-0.1, 0, 0.2], [0.0, 1.0, jerk_max_u])  #jerk_u가 0이 아니면, KONA_EV는 감속을 안함. over감속:upper를 +로 하면? 230930
-          jerk_l = interp(jerk, [-0.1, 0, 0.5], [jerk_max_l, 0.5, 0.5])  #jerk_l이 0.5가 아니면 가속도가 안올라감.
-        elif self.dynamicJerk == 2:
-          jerk_u = interp(a_diff, [-0.5, 0.1], [jerk_max, 0.2])
-          jerk_l = interp(a_diff, [-0.1, 0.5], [0.2, jerk_max])
-        else:
-          jerk_u = jerk_l = jerk_max
 
         can_sends.extend(hyundaican.create_acc_commands_mix_scc(self.CP, self.packer, CC.enabled, accel, jerk_u, jerk_l, int(self.frame / 2),
                                                       hud_control, set_speed_in_units, stopping, CC, CS, self.softHoldMode, speed_diff, self.car_fingerprint))