+++ selfdrive/ui/qt/network/networking.cc
@@ -12,6 +12,8 @@
 #include "selfdrive/ui/qt/widgets/controls.h"
 #include "selfdrive/ui/qt/widgets/scrollview.h"
 
+#include "common/params.h"
+
 static const int ICON_WIDTH = 49;
 
 // Networking functions
@@ -70,6 +72,13 @@ Networking::Networking(QWidget* parent, bool show_advanced) : QFrame(parent) {
     }
   )");
   main_layout->setCurrentWidget(wifiScreen);
+
+  if (Params().getBool("HotspotOnBoot")) {
+      printf("Enable TetheringConnection()\n");
+      //wifi->setTetheringEnabled(true);
+      an->toggleTethering(true);
+  }
+
 }
 
 void Networking::refresh() {
@@ -109,7 +118,7 @@ void Networking::hideEvent(QHideEvent *event) {
 }
 
 // AdvancedNetworking functions
-
+bool isOnBoot = true;
 AdvancedNetworking::AdvancedNetworking(QWidget* parent, WifiManager* wifi): QWidget(parent), wifi(wifi) {
 
   QVBoxLayout* main_layout = new QVBoxLayout(this);
@@ -125,7 +134,10 @@ AdvancedNetworking::AdvancedNetworking(QWidget* parent, WifiManager* wifi): QWid
 
   ListWidget *list = new ListWidget(this);
   // Enable tethering layout
-  tetheringToggle = new ToggleControl(tr("Enable Tethering"), "", "", wifi->isTetheringEnabled());
+  bool isTetheringEnabled = wifi->isTetheringEnabled();
+  if (isOnBoot && Params().getBool("HotspotOnBoot")) isTetheringEnabled = true;
+  isOnBoot = false;
+  tetheringToggle = new ToggleControl(tr("Enable Tethering"), "", "", isTetheringEnabled);
   list->addItem(tetheringToggle);
   QObject::connect(tetheringToggle, &ToggleControl::toggleFlipped, this, &AdvancedNetworking::toggleTethering);
 
@@ -192,6 +204,7 @@ AdvancedNetworking::AdvancedNetworking(QWidget* parent, WifiManager* wifi): QWid
 
   main_layout->addWidget(new ScrollView(list, this));
   main_layout->addStretch(1);
+
 }
 
 void AdvancedNetworking::refresh() {