+++ selfdrive/ui/qt/network/wifi_manager.cc
@@ -59,6 +59,7 @@ WifiManager::WifiManager(QObject *parent) : QObject(parent) {
   }
 
   timer.callOnTimeout(this, &WifiManager::requestScan);
+
 }
 
 void WifiManager::setup() {